SquirrelPAK AVP - Beta 1
================================================================================
SquirrelPAK AV - Almost Vanilla PLUS!
Change Log
================================================================================

This is the first open beta release!

Known Issues:
There is one known issue and that is with the options.txt file. There are issues with the mods I usually use so until we get that fix I will ship a options file. Before updating to new versions if you make changes to your options file make a backup copy first and replace it after updating.